Move shared test environment context under sub-package
