+++ server/src/test/java/io/spine/server/entity/AbstractEventSubscriberTest.java
@@ -29,21 +29,21 @@ import io.spine.core.EventId;
 import io.spine.core.UserId;
 import io.spine.server.BoundedContext;
 import io.spine.server.event.model.InsufficientVisibilityError;
-import io.spine.server.groups.FilteredStateSubscriber;
-import io.spine.server.groups.Group;
-import io.spine.server.groups.GroupId;
-import io.spine.server.groups.HiddenEntitySubscriber;
-import io.spine.server.groups.TestSubscriber;
-import io.spine.server.groups.WronglyDomesticSubscriber;
-import io.spine.server.groups.WronglyExternalSubscriber;
-import io.spine.server.organizations.Organization;
-import io.spine.server.organizations.OrganizationId;
+import io.spine.server.given.groups.FilteredStateSubscriber;
+import io.spine.server.given.groups.Group;
+import io.spine.server.given.groups.GroupId;
+import io.spine.server.given.groups.HiddenEntitySubscriber;
+import io.spine.server.given.groups.TestSubscriber;
+import io.spine.server.given.groups.WronglyDomesticSubscriber;
+import io.spine.server.given.groups.WronglyExternalSubscriber;
+import io.spine.server.given.organizations.Organization;
+import io.spine.server.given.organizations.OrganizationId;
 import io.spine.server.transport.memory.InMemoryTransportFactory;
 import io.spine.server.type.given.GivenEvent;
 import io.spine.system.server.DispatchedMessageId;
 import io.spine.system.server.EntityHistoryId;
-import io.spine.system.server.event.EntityStateChanged;
 import io.spine.system.server.SystemBoundedContexts;
+import io.spine.system.server.event.EntityStateChanged;
 import io.spine.type.TypeUrl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
