+++ server/src/test/java/io/spine/server/given/groups/WronglyDomesticSubscriber.java
@@ -18,21 +18,19 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
+package io.spine.server.given.groups;
 
-/**
- * A bounded context for testing entity state updates.
- *
- * <p>This package hosts a number of Protobuf definitions marked an {@code (entity)}-s and event
- * subscribers which receive updates of those entities.
- */
+import io.spine.core.Subscribe;
+import io.spine.server.event.AbstractEventSubscriber;
+import io.spine.server.given.organizations.Organization;
 
-@BoundedContext("Groups")
+import static org.junit.jupiter.api.Assertions.fail;
 
-@CheckReturnValue
-@ParametersAreNonnullByDefault
-package io.spine.server.groups;
+public class WronglyDomesticSubscriber extends AbstractEventSubscriber {
 
-import com.google.errorprone.annotations.CheckReturnValue;
-import io.spine.server.annotation.BoundedContext;
-
-import javax.annotation.ParametersAreNonnullByDefault;
+    @Subscribe // <-- Error here. Should be external.
+    void on(Organization organization) {
+        fail(WronglyDomesticSubscriber.class.getSimpleName() +
+                     " should not be able to receive external updates.");
+    }
+}
