+++ server/src/test/java/io/spine/server/given/groups/WronglyExternalSubscriber.java
@@ -18,19 +18,18 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-/**
- * A bounded context for testing entity state updates.
- *
- * <p>This package hosts a number of Protobuf definitions marked an {@code (entity)}-s.
- */
+package io.spine.server.given.groups;
 
-@BoundedContext("Organizations")
+import io.spine.core.Subscribe;
+import io.spine.server.event.AbstractEventSubscriber;
 
-@CheckReturnValue
-@ParametersAreNonnullByDefault
-package io.spine.server.organizations;
+import static org.junit.jupiter.api.Assertions.fail;
 
-import com.google.errorprone.annotations.CheckReturnValue;
-import io.spine.server.annotation.BoundedContext;
+public class WronglyExternalSubscriber extends AbstractEventSubscriber {
 
-import javax.annotation.ParametersAreNonnullByDefault;
+    @Subscribe(external = true) // <-- Error here. Should be domestic.
+    void on(Group group) {
+        fail(WronglyExternalSubscriber.class.getSimpleName() +
+                     " should not be able to receive domestic updates.");
+    }
+}
