+++ server/src/test/java/io/spine/server/given/organizations/package-info.java
@@ -18,18 +18,19 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.groups;
+/**
+ * A bounded context for testing entity state updates.
+ *
+ * <p>This package hosts a number of Protobuf definitions marked an {@code (entity)}-s.
+ */
 
-import io.spine.core.Subscribe;
-import io.spine.server.event.AbstractEventSubscriber;
+@BoundedContext("Organizations")
 
-import static org.junit.jupiter.api.Assertions.fail;
+@CheckReturnValue
+@ParametersAreNonnullByDefault
+package io.spine.server.given.organizations;
 
-public class WronglyExternalSubscriber extends AbstractEventSubscriber {
+import com.google.errorprone.annotations.CheckReturnValue;
+import io.spine.server.annotation.BoundedContext;
 
-    @Subscribe(external = true) // <-- Error here. Should be domestic.
-    void on(Group group) {
-        fail(WronglyExternalSubscriber.class.getSimpleName() +
-                     " should not be able to receive domestic updates.");
-    }
-}
+import javax.annotation.ParametersAreNonnullByDefault;
