+++ server/src/test/java/io/spine/server/projection/e2e/ProjectionEndToEndTest.java
@@ -30,15 +30,15 @@ import io.spine.core.EventContext;
 import io.spine.core.Events;
 import io.spine.core.UserId;
 import io.spine.server.BoundedContext;
-import io.spine.server.groups.Group;
-import io.spine.server.groups.GroupId;
-import io.spine.server.groups.GroupName;
-import io.spine.server.groups.GroupNameProjection;
-import io.spine.server.groups.GroupProjection;
-import io.spine.server.organizations.Organization;
-import io.spine.server.organizations.OrganizationEstablished;
-import io.spine.server.organizations.OrganizationId;
-import io.spine.server.organizations.OrganizationProjection;
+import io.spine.server.given.groups.Group;
+import io.spine.server.given.groups.GroupId;
+import io.spine.server.given.groups.GroupName;
+import io.spine.server.given.groups.GroupNameProjection;
+import io.spine.server.given.groups.GroupProjection;
+import io.spine.server.given.organizations.Organization;
+import io.spine.server.given.organizations.OrganizationEstablished;
+import io.spine.server.given.organizations.OrganizationId;
+import io.spine.server.given.organizations.OrganizationProjection;
 import io.spine.server.projection.given.EntitySubscriberProjection;
 import io.spine.server.projection.given.ProjectionRepositoryTestEnv.GivenEventMessage;
 import io.spine.server.projection.given.TestProjection;
