+++ server/src/test/java/io/spine/server/projection/given/ProjectionRepositoryTestEnv.java
@@ -24,8 +24,8 @@ import io.spine.core.EventContext;
 import io.spine.core.EventId;
 import io.spine.core.Subscribe;
 import io.spine.core.UserId;
-import io.spine.server.organizations.OrganizationEstablished;
-import io.spine.server.organizations.OrganizationId;
+import io.spine.server.given.organizations.OrganizationEstablished;
+import io.spine.server.given.organizations.OrganizationId;
 import io.spine.server.projection.Projection;
 import io.spine.server.projection.ProjectionRepository;
 import io.spine.server.type.MessageEnvelope;
