+++ server/src/test/proto/spine/test/entity/groups/group.proto
@@ -5,7 +5,7 @@ package spine.test.server.groups;
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.server.groups";
+option java_package = "io.spine.server.given.groups";
 option java_outer_classname = "GroupProto";
 option java_multiple_files = true;
