+++ server/src/test/proto/spine/test/entity/organizations/events.proto
@@ -5,7 +5,7 @@ package spine.test.server.organizations;
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.server.organizations";
+option java_package = "io.spine.server.given.organizations";
 option java_outer_classname = "EventsProto";
 option java_multiple_files = true;
